Split route tests so that one part is run in node and one in the browser
