+++ test/routes/change-password.server.test.ts
@@ -1,13 +1,10 @@
-import { screen, waitFor } from '@testing-library/react';
 import { fromPartial } from '@total-typescript/shoehorn';
-import { createRoutesStub } from 'react-router';
 import type { AuthHelper } from '../../app/auth/auth-helper.server';
-import ChangePassword, { action, loader } from '../../app/routes/change-password';
+import { action, loader } from '../../app/routes/change-password';
 import { INVALID_PASSWORD_FORMAT } from '../../app/routes/profile/change-password-action.server';
 import { PasswordMismatchError } from '../../app/users/PasswordMismatchError.server';
 import type { UsersService } from '../../app/users/UsersService.server';
 import { ValidationError } from '../../app/validation/ValidationError.server';
-import { renderWithEvents } from '../__helpers__/set-up-test';
 
 describe('change-password', () => {
   describe('loader', () => {
@@ -88,43 +85,6 @@ describe('change-password', () => {
       expect(result.init?.headers).toEqual(sessionCookie ? { 'Set-Cookie': sessionCookie } : undefined);
       expect(editUserTempPassword).toHaveBeenCalledOnce();
       expect(updateSession).toHaveBeenCalledOnce();
-    });
-  });
-
-  describe('<ChangePassword />', () => {
-    const setUp = async (error?: string) => {
-      const path = '/change-password';
-      const Stub = createRoutesStub([
-        {
-          path,
-          Component: ChangePassword,
-          HydrateFallback: () => null,
-          action: () => error ? { ok: false, error } : undefined,
-        },
-      ]);
-
-      const renderResult = renderWithEvents(<Stub initialEntries={[path]} />);
-      await screen.findByText(/^You need to change your temporary password/);
-
-      return renderResult;
-    };
-
-    it.each([
-      undefined,
-      'There was an error',
-    ])('shows error only if action response fails', async (error) => {
-      const { user } = await setUp(error);
-
-      // Send form so that the fetcher invokes the action
-      await user.type(screen.getByLabelText(/^New password/), 'aA123456!');
-      await user.type(screen.getByLabelText(/^Repeat password/), 'aA123456!');
-      await user.click(screen.getByRole('button', { name: 'Save' }));
-
-      if (error) {
-        await waitFor(() => expect(screen.getByTestId('error-container')).toHaveTextContent(error));
-      } else {
-        expect(screen.queryByTestId('error-container')).not.toBeInTheDocument();
-      }
     });
   });
 });
