+++ test/routes/servers/create-server.client.test.tsx
@@ -1,32 +1,10 @@
 import { screen, waitFor } from '@testing-library/react';
-import { fromPartial } from '@total-typescript/shoehorn';
 import { createRoutesStub } from 'react-router';
-import CreateServer, { action } from '../../../app/routes/servers/create-server';
-import type { ServersService } from '../../../app/servers/ServersService.server';
+import CreateServer from '../../../app/routes/servers/create-server';
 import { checkAccessibility } from '../../__helpers__/accessibility';
 import { renderWithEvents } from '../../__helpers__/set-up-test';
 
 describe('create-server', () => {
-  describe('action', () => {
-    const createServerForUser = vi.fn();
-    const serversService: ServersService = fromPartial({ createServerForUser });
-    const runAction = () => action(
-      fromPartial({
-        request: { formData: vi.fn().mockResolvedValue(new FormData()) },
-        context: { get: vi.fn().mockReturnValue({ publicId: '123' }) },
-      }),
-      serversService,
-    );
-
-    it('redirects to servers list after creating a server', async () => {
-      const resp = await runAction();
-
-      expect(resp.status).toEqual(302);
-      expect(resp.headers.get('Location')).toEqual('/manage-servers/1');
-      expect(createServerForUser).toHaveBeenLastCalledWith('123', new FormData());
-    });
-  });
-
   describe('<CreateServer />', () => {
     const setUp = async () => {
       const path = '/manage-servers/create';
