+++ test/routes/settings.server.test.ts
@@ -1,10 +1,8 @@
 import type { Settings } from '@shlinkio/shlink-web-component/settings';
-import { render, screen, waitFor } from '@testing-library/react';
 import { fromPartial } from '@total-typescript/shoehorn';
 import type { ActionFunctionArgs } from 'react-router';
-import { createRoutesStub } from 'react-router';
 import type { AuthHelper } from '../../app/auth/auth-helper.server';
-import SettingsComp, { action as settingsAction, loader } from '../../app/routes/settings';
+import { action as settingsAction, loader } from '../../app/routes/settings';
 import type { SettingsService } from '../../app/settings/SettingsService.server';
 
 describe('settings', () => {
@@ -54,29 +52,6 @@ describe('settings', () => {
 
       expect(getSession).toHaveBeenCalledWith(request);
       expect(saveUserSettings).toHaveBeenCalledWith('1', {});
-    });
-  });
-
-  // FIXME Skipping for now, as the test always resolves a 404
-  describe('<Settings />', () => {
-    const setUp = () => {
-      const Stub = createRoutesStub([
-        {
-          path: '/settings/*',
-          Component: SettingsComp,
-          HydrateFallback: () => null,
-          loader: () => ({}),
-          action: () => ({}),
-        },
-      ]);
-      return render(<Stub initialEntries={['/settings/general']} />);
-    };
-
-    it('renders settings component', async () => {
-      setUp();
-
-      await waitFor(() => expect(screen.getByRole('heading', { name: 'User interface' })).toBeInTheDocument());
-      expect(screen.getByRole('heading', { name: 'Real-time updates' })).toBeInTheDocument();
     });
   });
 });
